add support/client.ts
