+++ explorer/src/__tests__/server/realtime.test.ts
@@ -3,42 +3,24 @@ import { Connection } from 'typeorm'
 import WebSocket from 'ws'
 import { getDb } from '../../database'
 import { ChainlinkNode, createChainlinkNode } from '../../entity/ChainlinkNode'
-import { DEFAULT_TEST_PORT, start, stop } from '../../support/server'
+import { JobRun } from '../../entity/JobRun'
+import { start, stop } from '../../support/server'
 import { clearDb } from '../testdatabase'
+import { NORMAL_CLOSE } from '../../utils/constants'
 import {
-  ACCESS_KEY_HEADER,
-  NORMAL_CLOSE,
-  SECRET_HEADER,
-} from '../../utils/constants'
-
-const ENDPOINT = `ws://localhost:${DEFAULT_TEST_PORT}`
-
-const newChainlinkNode = (
-  url: string,
-  accessKey: string,
-  secret: string,
-): Promise<WebSocket> => {
-  const ws = new WebSocket(ENDPOINT, {
-    headers: {
-      [ACCESS_KEY_HEADER]: accessKey,
-      [SECRET_HEADER]: secret,
-    },
-  })
-
-  return new Promise((resolve: (arg0: WebSocket) => void, reject) => {
-    ws.on('error', (error: Error) => {
-      reject(error)
-    })
-
-    ws.on('open', () => resolve(ws))
-  })
-}
+  ENDPOINT,
+  createRPCRequest,
+  newChainlinkNode,
+} from '../../support/client'
 
 describe('realtime', () => {
   let server: Server
   let db: Connection
   let chainlinkNode: ChainlinkNode
   let secret: string
+
+  const newAuthenticatedNode = async () =>
+    newChainlinkNode(ENDPOINT, chainlinkNode.accessKey, secret)
 
   beforeAll(async () => {
     server = await start()
@@ -54,6 +36,75 @@ describe('realtime', () => {
   })
 
   afterAll(done => stop(server, done))
+
+  describe('when sending messages in JSON-RPC format', () => {
+    it('rejects non-existing methods with code -32601', async (done: any) => {
+      expect.assertions(2)
+
+      const ws = await newAuthenticatedNode()
+      const request = createRPCRequest('doesNotExist')
+      ws.send(JSON.stringify(request))
+
+      ws.on('message', async (data: any) => {
+        const response = JSON.parse(data)
+        expect(response.error.code).toEqual(-32601)
+
+        const count = await db.manager.count(JobRun)
+        expect(count).toEqual(0)
+
+        ws.close()
+        done()
+      })
+    })
+
+    // this test depends on the presence of "jsonrpc" in the message
+    // otherwise, the server will attempt to process the message as a
+    // legacy message and will respond with { status: 422 }.
+    // This test will be more appropriate once the legacy format is removed.
+    it('rejects malformed json with code -32700', async (done: any) => {
+      expect.assertions(2)
+
+      const ws = await newAuthenticatedNode()
+      const request = 'jsonrpc invalid'
+      ws.send(request)
+
+      ws.on('message', async (data: any) => {
+        const response = JSON.parse(data)
+        expect(response.error.code).toEqual(-32700)
+
+        const count = await db.manager.count(JobRun)
+        expect(count).toEqual(0)
+
+        ws.close()
+        done()
+      })
+    })
+
+    it('rejects invalid rpc requests with code -32600', async (done: any) => {
+      expect.assertions(2)
+
+      const ws = await newAuthenticatedNode()
+
+      const request = {
+        jsonrpc: '2.0',
+        function: 'foo',
+        id: 1,
+      }
+
+      ws.send(JSON.stringify(request))
+
+      ws.on('message', async (data: any) => {
+        const response = JSON.parse(data)
+        expect(response.error.code).toEqual(-32600)
+
+        const count = await db.manager.count(JobRun)
+        expect(count).toEqual(0)
+
+        ws.close()
+        done()
+      })
+    })
+  })
 
   it('rejects invalid authentication', async (done: any) => {
     expect.assertions(1)
